requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.10
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
asyncio-mqtt==0.16.2
attrs==25.3.0
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
decorator==5.2.1
Deprecated==1.2.18
discord-webhook==1.4.1
executing==2.2.0
frozenlist==1.6.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-adk
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
loguru==0.7.3
matplotlib-inline==0.1.7
multidict==6.4.4
mypy==1.16.0
mypy_extensions==1.1.0
packaging==25.0
paho-mqtt==2.1.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
stack-data==0.6.3
tenacity==9.1.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.20.0
